groovebox: migrate onto the Core 0 input pipeline (#109)
Drums dispatch as plain Note (set_note() branches on channel) and
pattern-select moves from a NOTE_ON-shaped message to Program Change
under CONFIGURATION, matching every other migrated module's preset-select
convention. With MIDI_CLOCK added to midi_controller_process_generic(),
groovebox's midi_controller_process() collapses to the same one-line
call every other migrated module uses -- no more forked process loop.

midi_dispatch_strike() and patterns.h's SEQ_PATTERN_BASE_NOTE, both left
behind by the remap, are removed as dead code.
